fix wrong type of i in print method

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void print(Object... args){
-        for (int i=0; i< args.length; i++) {
+        for (byte i=0; i< args.length; i++) {
             System.out.println(args[i]);
         }
         System.out.println();
